Move explores and views to directory
This should make it easier to move around this repo.
